# Improve performance of replace repeated values with nan