--- a/test/unit/data_classes/test_split_function.py
+++ b/test/unit/data_classes/test_split_function.py
@@ -2,9 +2,10 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 
-import unittest
-import json
 import copy
+import json
+import unittest
+
 from openstef.data_classes.split_function import SplitFuncDataClass