--- a/test/unit/feature_engineering/test_apply_features.py
+++ b/test/unit/feature_engineering/test_apply_features.py
@@ -137,18 +137,14 @@ class TestApplyFeaturesModule(BaseTestCase):
         # Skip first row, since T-30min not available for first row
         self.assertTrue(
             input_data_with_features.loc[horizon == 47, ["APX", "T-30min"]]
-            .iloc[
-                1:,
-            ]
+            .iloc[1:,]
             .isna()
             .all()
             .all()
         )
         self.assertFalse(
             input_data_with_features.loc[horizon == 0.25, ["APX", "T-30min"]]
-            .iloc[
-                1:,
-            ]
+            .iloc[1:,]
             .isna()
             .any()
             .any()