Plus run isort and black

New function to replace nan has improved performance from:
725 ms ± 111 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
to:
2.41 ms ± 164 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)