@Davidswinkels 2022-07-06
Kept the first 24 flatline values :)